Extract list component
