+++ public/app/core/components/NestedFolderPicker/types.ts
@@ -1,0 +1,10 @@
+export interface RootFolder {
+  title: string;
+  id?: number;
+  uid: string;
+}
+
+export type RootFolderWithUiState = RootFolder & {
+  level: number;
+  expanded: boolean;
+};
